% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_tables}
\alias{get_tables}
\title{get_tables}
\usage{
get_tables(wikipediaurl)
}
\arguments{
\item{wikipediaurl}{A url on Wikipedia, including https://}
}
\value{
A list of tibbles
}
\description{
A function that grabs all html tables found on a url and puts
each table in a tibble, and all tibbles in a list.
}
\examples{
list_of_tibbles <- get_tables("https://en.wikipedia.org/wiki/Kentucky_Derby")
my_tibble <- list_of_tibbles[[2]]

}
